+++ src/main/java/model/algorithms/AlgorithmStrategy.java
@@ -2,5 +2,5 @@ package model.algorithms;
 
 public interface AlgorithmStrategy {
 
-    void solve(int[][] board);
+    void solve(byte[][] board);
 }
